fix: move license banner to seperate file
