+++ frontend/src/component/banners/externalBanners/ExternalBanners.tsx
@@ -5,8 +5,7 @@ import { useVariant } from 'hooks/useVariant';
 import { IBanner } from 'interfaces/banner';
 
 export const ExternalBanners = () => {
-    const { uiConfig, isEnterprise } = useUiConfig();
-    const licenseInfo = useLicenseCheck();
+    const { uiConfig } = useUiConfig();
 
     const bannerVariantFromMessageBannerFlag = useVariant<IBanner | IBanner[]>(
         uiConfig.flags.messageBanner,
@@ -21,24 +20,6 @@ export const ExternalBanners = () => {
     const banners: IBanner[] = Array.isArray(bannerVariant)
         ? bannerVariant
         : [bannerVariant];
-
-    // Only for enterprise
-    if (isEnterprise()) {
-        if (
-            licenseInfo &&
-            !licenseInfo.isValid &&
-            !licenseInfo.loading &&
-            !licenseInfo.error
-        ) {
-            banners.push({
-                message:
-                    licenseInfo.message ||
-                    'You have an invalid Unleash license.',
-                variant: 'error',
-                sticky: true,
-            });
-        }
-    }
 
     return (
         <>
